+++ src/main/java/de/mirb/project/miftp/fs/InMemoryFsView.java
@@ -184,6 +184,7 @@ public class InMemoryFsView implements FileSystemView {
 
   @Override
   public void dispose() {
-
+    LOG.debug("Dispose FSView");
+    workingDir = homeDir;
   }
 }