+++ src/main/java/de/mirb/project/miftp/fs/InMemoryFtpFile.java
@@ -1,7 +1,5 @@
 package de.mirb.project.miftp.fs;
 
-import org.apache.ftpserver.ftplet.User;
-
 import java.io.ByteArrayInputStream;
 import java.io.InputStream;
 import java.io.OutputStream;